fix: don’t crash when an image is missing in the EPUB
When an image resource is missing in the EPUB, the copy
operation is ignored and a warning message is logged.

Closes #44

+++ src/report/report.js
@@ -89,9 +89,17 @@ module.exports = class Report {
               const fromPath = img.path;
               const toPath = path.join(outdir, 'data', img.src);
               delete img.path;
-              return fs.copy(fromPath, toPath, {
-                overwrite: false,
-              });
+              return fs.pathExists(fromPath)
+                .then((exists) => {
+                  if (exists) {
+                    return fs.copy(fromPath, toPath, {
+                      overwrite: false,
+                    });
+                  } else {
+                    winston.warn(`Couldn’t copy resource '${img.src}'`);
+                    return Promise.resolve();
+                  }
+                });
             });
           }
         });
